Refactor: use floats instead of intergers

+++ lib/budget_tracking_tool/budgets/budget.ex
@@ -3,7 +3,7 @@ defmodule BudgetTrackingTool.Budgets.Budget do
   import Ecto.Changeset
 
   schema "budgets" do
-    field :amount, :integer
+    field :amount, :float
     field :date, :date
 
     belongs_to :org, BudgetTrackingTool.Accounts.Org
